removed values copies in the mask applier
The image is modified in the operator|= so we don't have to write
multiple copies on it. For instance, the mask applier creates a copy of
the image to compute the new values of the pixels. At the end, we do not
"sync" the copy and the actual image yet. Instead we return the copy
that is the new image that is modified threw the algorithms. One we have
the final result, we update the image so we do it only once.